Extra Functionality Commit #2
The project now writes the name of the state in the state (unless it is a small space, then it is found relatively close to the state area). I adjusted each state name positioning manually in relation to the averaged coordinate because the coordinates distort some of the state's average coordinates.